// types/types.d.ts
export type SvgProps = {
  width?: number;
  height?: number;
  color?: string;
};

export type LanguageProps =
  | "HTML"
  | "CSS"
  | "JS"
  | "ReactJs"
  | "NextJs"
  | "NodeJs"
  | "RemixJs"
  | "Python"
  | "Bootstrap"
  | "Tailwind"
  | "Git"
  | "Typescript"
  | "Sql"
  | "PostgreSql"
  | "MongoDb"
  | "Prisma"
  | "ExpressJs";